+++ raspd/quadcopter.h
@@ -3,7 +3,7 @@
 
 int pidctrl_init(int front, int rear, int left, int right,
                 long (*get_altitude)(unsigned long *timestamp),
-                long angle[], long rate[], long alti[]);
+                double angle[], double rate[], double alti[]);
 void pidctrl_exit(void);
 
 #endif /* __QUADCOPTER_H__ */
